requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Cors>=5.0.0,<6.0.0
Flask-SQLAlchemy>=3.1.0,<4.0.0
joblib>=1.4.0,<2.0.0
pandas>=2.2.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
scikit-learn>=1.6.0,<2.0.0
